Move predicates reader to utils
