+++ spark/submit.py
@@ -1,23 +1,9 @@
 from spark.session import SharedSparkSession
 from spark.job import SparkJob
+from spark.utils import read_predicates_from_csv
 from pathlib import Path
-import csv
-
-
-def read_predicates_from_csv(path: Path) -> list[str]:
-    predicates = []
-    with open(path.resolve().as_posix(), mode="r") as file:
-        csv_reader = csv.reader(file)
-        next(csv_reader)
-        for row in csv_reader:
-            start, end, _ = row
-            predicate = f"parid >= '{start}' AND parid <= '{end}'"
-            predicates.append(predicate)
-    return predicates
-
 
 predicates = read_predicates_from_csv(Path("/tmp/spark/predicates.csv"))
-
 session = SharedSparkSession(app_name="iasworld")
 
 addn = SparkJob(
